requirements.txt
openai>=1.40.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
# Vector store.
chromadb>=0.5.0
sentence-transformers>=2.7.0
jupyter>=1.0.0
ipykernel>=6.29.0
